# openapi.yml
openapi: 3.0.0
info:
  description: Backend API for the Stanford digital repository
  version: 1.0.0
  title: DOR Services API
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://dor-services-{env}.stanford.edu"
    description: Production service
    variables:
      env:
        default: prod
  - url: "https://dor-services-{env}.stanford.edu"
    description: Staging service
    variables:
      env:
        default: stage
...
paths:
...
  "/v1/objects/{id}/publish":
    post:
      tags:
        - objects
      summary: Publish object to PURL
      description: ""
      operationId: "objects#publish"
      responses:
        "201":
          description: Publishing action started
          headers:
            location:
              description: the status of the action is found at this URI
              schema:
                type: string
                format: uri
      parameters:
        - name: id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
        - name: lane-id
          in: query
          schema:
            $ref: "#/components/schemas/LaneId"
  "/v1/objects/{id}/accession":
    post:
      tags:
        - objects
      summary: Start an assembly workflow, optionally opening a version
      description: ""
      operationId: "objects#accession"
      responses:
        "201":
          description: Workflow started
        "409":
          description: Unable to open version.
      parameters:
        - name: id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
        - name: workflow
          in: query
          description: which workflow should be started
          required: false
          schema:
            type: string
            default: assemblyWF
            enum:
              - assemblyWF
              - gisAssemblyWF
        - name: description
          in: query
          description: the description of the version change (if versioning is needed)
          required: true
          schema:
            type: string
          example: "Re-accessioning this object"
        - name: opening_user_name
          in: query
          description: the username of the person creating the version (if versioning is needed)
          required: false
          schema:
            type: string
          example: "some_sunetid"
  "/v1/objects/{id}/reindex":
    post:
      tags:
        - integrations
      summary: Reindex the object in Solr
      description: ""
      operationId: "objects#reindex"
      responses:
        "204":
          description: OK
      parameters:
        - name: id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
  "/v1/objects/{id}/indexable":
    post:
      tags:
        - integrations
      summary: Validate the object for indexing in Solr without actually indexing it
      description: ""
      operationId: "objects#indexable"
      parameters:
        - name: id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
      requestBody:
        required: true
        content:
          application/json:
            # Note that additionalProperties needs to be set to true to allow id parameter.
            schema:
              oneOf:
                - $ref: "#/components/schemas/DRO"
                - $ref: "#/components/schemas/Collection"
                - $ref: "#/components/schemas/AdminPolicy"
      responses:
        "204":
          description: OK
        "422":
          description: Index validation has failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
...
  "/v1/objects/{object_id}/administrative_tags/{id}":
    delete:
      tags:
        - tags
      summary: Delete an administrative tag for an object
      description: ""
      operationId: "administrative_tags#destroy"
      responses:
        "204":
          description: Deleted
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Object or tag not found
          content:
            text/plain:
              schema:
                type: string
      parameters:
        - name: object_id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
        - name: id
          in: path
          description: ID of tag to remove
          required: true
          schema:
            $ref: "#/components/schemas/AdministrativeTagEscaped"
    put:
      tags:
        - tags
      summary: Update one administrative tag
      description: ""
      operationId: "administrative_tags#update (PUT)"
      responses:
        "204":
          description: Updated
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Object or tag not found
          content:
            text/plain:
              schema:
                type: string
        "409":
          description: Request would result in a duplicate tag
          content:
            text/plain:
              schema:
                type: string
      parameters:
        - name: object_id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
        - name: id
          in: path
          description: ID of tag to replace
          required: true
          schema:
            $ref: "#/components/schemas/AdministrativeTagEscaped"
      requestBody:
        description: A replacement administrative tag
        content:
          application/json:
            schema:
              type: object
              required:
                - administrative_tag
              properties:
                administrative_tag:
                  $ref: "#/components/schemas/AdministrativeTag"
    patch:
      tags:
        - tags
      summary: Update one administrative tag
      description: ""
      operationId: "administrative_tags#update (PATCH)"
      responses:
        "204":
          description: Updated
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Object or tag not found
        "409":
          description: Request would result in a duplicate tag
      parameters:
        - name: object_id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
        - name: id
          in: path
          description: ID of tag to replace
          required: true
          schema:
            $ref: "#/components/schemas/AdministrativeTagEscaped"
      requestBody:
        description: A replacement administrative tag
        content:
          application/json:
            schema:
              type: object
              required:
                - administrative_tag
              properties:
                administrative_tag:
                  $ref: "#/components/schemas/AdministrativeTag"
...
  /v1/objects/find_all:
    post:
      tags:
        - objects
      summary: Finds multiple objects
      operationId: "objects#find_all"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                externalIdentifiers:
                  type: array
                  items:
                    $ref: "#/components/schemas/Druid"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: "#/components/schemas/DRO"
                    - $ref: "#/components/schemas/Collection"
                    - $ref: "#/components/schemas/AdminPolicy"
...
  "/v1/objects/{id}":
    patch:
      tags:
        - objects
      summary: Update object metadata
      description: |-
        Replace the object metadata with the provided metadata.

        The returned Cocina object may differ from the provided Cocina object due to normalization.

        Since we don't presently have a full mapping between cocina JSON and MODS,
        only the title can be updated in the descriptive metadata.
      operationId: "objects#update"
      parameters:
        - name: id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
        - name: event_description
          in: query
          description: A description of the change being made to the object
          required: false
          schema:
            type: string
        - name: user_name
          in: query
          description: The sunetid of the person making the change
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            # Note that additionalProperties needs to be set to true to allow id parameter.
            schema:
              oneOf:
                - $ref: "#/components/schemas/DRO"
                - $ref: "#/components/schemas/Collection"
                - $ref: "#/components/schemas/AdminPolicy"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/DRO"
                  - $ref: "#/components/schemas/Collection"
                  - $ref: "#/components/schemas/AdminPolicy"
          headers:
            ETag:
              schema:
                type: string
            Last-Modified:
              schema:
                type: string
            X-Created-At:
              schema:
                type: string
        "400":
          description: Invalid DOR parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Object not found in DOR
        "409":
          description: Object with that sourceId already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "412":
          description: Roundtrip validation has failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
    get:
      tags:
        - objects
      summary: Retrieve the object COCINA metadata
      description: "Returns a JSON representation (not yet complete) of an object, collection or admin policy."
      operationId: "objects#show"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/DRO"
                  - $ref: "#/components/schemas/Collection"
                  - $ref: "#/components/schemas/AdminPolicy"
          headers:
            ETag:
              schema:
                type: string
            Last-Modified:
              schema:
                type: string
            X-Created-At:
              schema:
                type: string
        "304":
          description: No change. The client provided an ETag that matches the current state of the item.
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
    head:
      tags:
        - objects
      summary: Retrieve the head for object COCINA metadata
      operationId: "objects#head"
      responses:
        "200":
          description: OK
          headers:
            ETag:
              schema:
                type: string
            Last-Modified:
              schema:
                type: string
            X-Created-At:
              schema:
                type: string
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
    delete:
      tags:
        - objects
      summary: Destroys the object
      description: ""
      operationId: "objects#destroy"
      responses:
        "201":
          description: Deleted
        "400":
          description: Bad request
        "404":
          description: Object or tag not found
        "500":
          description: Internal Server Error
      parameters:
        - name: id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
        - name: user_name
          in: query
          description: Username of user requesting delete
          required: false
          schema:
            type: string
...
components:
  schemas:
    Access:
      type: object
      oneOf:
        # Being first, makes DarkAccess the default.
        - $ref: "#/components/schemas/DarkAccess"
        - $ref: "#/components/schemas/CitationOnlyAccess"
        - $ref: "#/components/schemas/ControlledDigitalLendingAccess"
        - $ref: "#/components/schemas/LocationBasedAccess"
        - $ref: "#/components/schemas/LocationBasedDownloadAccess"
        - $ref: "#/components/schemas/StanfordAccess"
        - $ref: "#/components/schemas/WorldAccess"
...
    AdminPolicy:
      type: object
      additionalProperties: true
      properties:
        cocinaVersion:
          $ref: "#/components/schemas/CocinaVersion"
        type:
          type: string
          enum:
            - "https://cocina.sul.stanford.edu/models/admin_policy"
        externalIdentifier:
          $ref: "#/components/schemas/Druid"
        label:
          type: string
        version:
          type: integer
        administrative:
          $ref: "#/components/schemas/AdminPolicyAdministrative"
        description:
          $ref: "#/components/schemas/Description"
      required:
        - cocinaVersion
        - administrative
        - externalIdentifier
        - label
        - type
        - version
    Administrative:
      type: object
      properties:
        hasAdminPolicy:
          $ref: "#/components/schemas/Druid"
      required:
        - hasAdminPolicy
...
    AdminPolicyAdministrative:
      description: Administrative properties for an AdminPolicy
      type: object
      additionalProperties: false
      properties:
        accessTemplate:
          $ref: "#/components/schemas/AdminPolicyAccessTemplate"
        registrationWorkflow:
          description: When you register an item with this admin policy, these are the workflows that are available.
          type: array
          items:
            type: string
        disseminationWorkflow:
          description: An additional workflow to start for objects managed by this admin policy once the end-accession workflow step is complete
          example: wasCrawlPreassemblyWF
          type: string
        collectionsForRegistration:
          description: When you register an item with this admin policy, these are the collections that are available.
          type: array
          items:
            type: string
        hasAdminPolicy:
          $ref: "#/components/schemas/Druid"
        hasAgreement:
          $ref: "#/components/schemas/Druid"
        roles:
          description: The access roles conferred by this AdminPolicy (used by Argo)
          type: array
          items:
            $ref: "#/components/schemas/AccessRole"
      required:
        - hasAdminPolicy
        - hasAgreement
        - accessTemplate
...
    CocinaVersion:
      description: The version of Cocina with which this object conforms.
      type: string
      pattern: '^\d+\.\d+\.\d+$'
      example: "1.2.3"
    Collection:
      description: A group of Digital Repository Objects that indicate some type of conceptual grouping within the domain that is worth reusing across the system.
      type: object
      additionalProperties: true
      properties:
        cocinaVersion:
          $ref: "#/components/schemas/CocinaVersion"
        type:
          description: The content type of the Collection. Selected from an established set of values.
          type: string
          enum:
            - "https://cocina.sul.stanford.edu/models/collection"
            - "https://cocina.sul.stanford.edu/models/curated-collection"
            - "https://cocina.sul.stanford.edu/models/user-collection"
            - "https://cocina.sul.stanford.edu/models/exhibit"
            - "https://cocina.sul.stanford.edu/models/series"
        externalIdentifier:
          $ref: "#/components/schemas/Druid"
        label:
          description: Primary processing label (can be same as title) for a Collection.
          type: string
        version:
          description: Version for the Collection within SDR.
          type: integer
        access:
          $ref: "#/components/schemas/CollectionAccess"
        administrative:
          $ref: "#/components/schemas/Administrative"
        description:
          $ref: "#/components/schemas/Description"
        identification:
          $ref: "#/components/schemas/CollectionIdentification"
      required:
        - cocinaVersion
        - description
        - externalIdentifier
        - label
        - type
        - version
        - access
        - administrative
        - identification
    CollectionAccess:
      description: Access metadata for collections
      type: object
      additionalProperties: false
      properties:
        view:
          description: Access level
          type: string
          enum:
            - "world"
            - "dark"
          default: "dark"
        copyright:
          $ref: "#/components/schemas/Copyright"
        useAndReproductionStatement:
          $ref: "#/components/schemas/UseAndReproductionStatement"
        license:
          $ref: "#/components/schemas/License"
    CollectionIdentification:
      type: object
      additionalProperties: false
      properties:
        catalogLinks:
          type: array
          items:
            $ref: "#/components/schemas/CatalogLink"
        sourceId:
          $ref: "#/components/schemas/SourceId"
...
    Copyright:
      description: The human readable copyright statement that applies
      example: Copyright World Trade Organization
      type: string
      nullable: true
...
    DRO:
      description: Domain-defined abstraction of a 'work'. Digital Repository Objects' abstraction is describable for our domain’s purposes, i.e. for management needs within our system.
      type: object
      additionalProperties: true
      properties:
        cocinaVersion:
          $ref: "#/components/schemas/CocinaVersion"
        type:
          description: The content type of the DRO. Selected from an established set of values.
          type: string
          enum:
            - "https://cocina.sul.stanford.edu/models/object"
            - "https://cocina.sul.stanford.edu/models/3d"
            - "https://cocina.sul.stanford.edu/models/agreement"
            - "https://cocina.sul.stanford.edu/models/book"
            - "https://cocina.sul.stanford.edu/models/document"
            - "https://cocina.sul.stanford.edu/models/geo"
            - "https://cocina.sul.stanford.edu/models/image"
            - "https://cocina.sul.stanford.edu/models/page"
            - "https://cocina.sul.stanford.edu/models/photograph"
            - "https://cocina.sul.stanford.edu/models/manuscript"
            - "https://cocina.sul.stanford.edu/models/map"
            - "https://cocina.sul.stanford.edu/models/media"
            - "https://cocina.sul.stanford.edu/models/track"
            - "https://cocina.sul.stanford.edu/models/webarchive-binary"
            - "https://cocina.sul.stanford.edu/models/webarchive-seed"
        externalIdentifier:
          $ref: "#/components/schemas/Druid"
        label:
          description: Primary processing label (can be same as title) for a DRO.
          type: string
        version:
          description: Version for the DRO within SDR.
          type: integer
        access:
          $ref: "#/components/schemas/DROAccess"
        administrative:
          $ref: "#/components/schemas/Administrative"
        description:
          $ref: "#/components/schemas/Description"
        identification:
          $ref: "#/components/schemas/Identification"
        structural:
          $ref: "#/components/schemas/DROStructural"
        geographic:
          $ref: "#/components/schemas/Geographic"
      required:
        - cocinaVersion
        - access
        - administrative
        - description
        - externalIdentifier
        - label
        - type
        - version
        - identification
        - structural
    DROAccess:
      type: object
      allOf:
        - $ref: "#/components/schemas/Access"
        - type: object
          properties:
            copyright:
              $ref: "#/components/schemas/Copyright"
            embargo:
              $ref: "#/components/schemas/Embargo"
            useAndReproductionStatement:
              $ref: "#/components/schemas/UseAndReproductionStatement"
            license:
              $ref: "#/components/schemas/License"
    DROStructural:
      description: Structural metadata
      type: object
      additionalProperties: false
      properties:
        contains:
          description: Filesets that contain the digital representations (Files)
          type: array
          items:
            $ref: "#/components/schemas/FileSet"
        hasMemberOrders:
          description: Provided sequences or orderings of members, including some metadata about each sequence (i.e. sequence label, sequence type, if the sequence is primary, etc.).
          type: array
          items:
            $ref: "#/components/schemas/Sequence"
        isMemberOf:
          description: Collections that this DRO is a member of
          type: array
          items:
            $ref: "#/components/schemas/Druid"
...
    Description:
      type: object
      allOf:
        - $ref: "#/components/schemas/RequestDescription"
        - type: object
          properties:
            purl:
              $ref: "#/components/schemas/Purl"
          required:
            - purl
...
    Druid:
      type: string
      pattern: "^druid:[b-df-hjkmnp-tv-z]{2}[0-9]{3}[b-df-hjkmnp-tv-z]{2}[0-9]{4}$"
      example: "druid:bc123df4567"
    Embargo:
      type: object
      allOf:
        - $ref: "#/components/schemas/Access"
        - type: object
          properties:
            releaseDate:
              description: Date when the Collection is released from an embargo.
              type: string
              format: date-time
              example: "2029-06-22T07:00:00.000+00:00"
            useAndReproductionStatement:
              $ref: "#/components/schemas/UseAndReproductionStatement"
          required:
            - releaseDate
...
    Geographic:
      description: Geographic metadata
      type: object
      additionalProperties: false
      properties:
        iso19139:
          description: Geographic ISO 19139 XML metadata
          type: string
      required:
        - iso19139
    Identification:
      type: object
      additionalProperties: false
      properties:
        barcode:
          $ref: "#/components/schemas/Barcode"
        catalogLinks:
          type: array
          items:
            $ref: "#/components/schemas/CatalogLink"
        doi:
          $ref: "#/components/schemas/DOI"
        sourceId:
          $ref: "#/components/schemas/SourceId"
      required:
        - sourceId
...
    AdministrativeTag:
      type: string
      pattern: "^.+( : .+)+$"
      example: "Foo : Bar : Baz"
    AdministrativeTagEscaped:
      type: string
      # allow CGI-escaped and bare colons to allow for differences between local
      # environments and deployed environments that use Apache & Passenger
      pattern: '^.+(\+(?:%3A|:)\+.+)+$'
      example: "Foo+%3A+Bar+%3A+Baz"
    LaneId:
      description: Lane for prioritizing the work
      type: string
      enum:
        - "default"
        - "low"
      default: "default"
...
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"
    Error:
      type: object
      properties:
        title:
          type: string
          description: "a short, human-readable summary of the problem that SHOULD NOT change from occurrence to occurrence of the problem."
          example: Invalid Attribute
        detail:
          type: string
          description: a human-readable explanation specific to this occurrence of the problem.
          example: Title must contain at least three characters.
        meta:
          type: object
          description: used to include non-standard meta-information
          example: { backtrace: ["where things went wrong"] }
        source:
          type: object
          properties:
            pointer:
              type: string
              example: /data/attributes/title
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []